# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Salón de Belleza",
    "description": "API RESTful para gestión de salón de estética: citas, clientes, servicios, inventario, proveedores, compras, ventas, finanzas y cortes de caja.",
    "version": "2.0.0"
  },
  "servers": [
    { "url": "http://localhost:5000", "description": "Desarrollo local" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "Mensaje de error" }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": { "type": "string", "example": "admin" },
          "password": { "type": "string", "example": "123456" }
        },
        "required": ["username", "password"]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "usu_username": { "type": "string" },
          "usu_password": { "type": "string" },
          "usu_email": { "type": "string" }
        },
        "required": ["usu_username", "usu_password"]
      },
      "ChangePasswordRequest": {
        "type": "object",
        "properties": {
          "current_password": { "type": "string" },
          "new_password": { "type": "string" }
        },
        "required": ["current_password", "new_password"]
      },
      "Usuario": {
        "type": "object",
        "properties": {
          "usu_id": { "type": "integer" },
          "usu_username": { "type": "string" },
          "usu_email": { "type": "string" },
          "usu_estado": { "type": "string", "enum": ["activo", "inactivo"] }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Iniciar sesión",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } } },
        "responses": {
          "200": { "description": "Token JWT retornado" },
          "401": { "description": "Credenciales inválidas", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Registrar nuevo usuario",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } } } },
        "responses": {
          "201": { "description": "Usuario registrado" },
          "400": { "description": "Error de validación", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Autenticación"],
        "summary": "Obtener información del usuario autenticado",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Datos del usuario", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Usuario" } } } },
          "401": { "description": "No autorizado" }
        }
      }
    },
    "/api/auth/change-password": {
      "put": {
        "tags": ["Autenticación"],
        "summary": "Cambiar contraseña",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChangePasswordRequest" } } } },
        "responses": {
          "200": { "description": "Contraseña actualizada" },
          "400": { "description": "Error de validación", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}